feat(LINGUIST-66): Implemented streak update upon login and changed project artifact to bmvp
