+++ backend/.gitignore
@@ -31,5 +31,4 @@ build/
 
 ### VS Code ###
 .vscode/
-
 application.properties
