--- backend/src/main/java/app/linguistai/LinguistaiApplication.java
@@ -1,13 +1,0 @@
-package app.linguistai;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LinguistaiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LinguistaiApplication.class, args);
-	}
-
-}
