+++ backend/src/main/java/app/linguistai/bmvp/LinguistaiApplication.java
@@ -1,0 +1,13 @@
+package app.linguistai.bmvp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LinguistaiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LinguistaiApplication.class, args);
+	}
+
+}
