+++ backend/src/main/java/app/linguistai/bmvp/exception/ExceptionLogger.java
@@ -1,0 +1,18 @@
+package app.linguistai.bmvp.exception;
+
+import lombok.Data;
+
+@Data
+public class ExceptionLogger {
+    public static String log(Exception e) {
+        String log = "";
+
+        if (e.getMessage().equals(e.getLocalizedMessage())) {
+            log += e.getLocalizedMessage();
+        } else {
+            log += e.getLocalizedMessage() + " : " + e.getMessage();
+        }
+
+        return log;
+    }
+}
