+++ backend/src/main/java/app/linguistai/bmvp/exception/NotFoundException.java
@@ -1,0 +1,11 @@
+package app.linguistai.bmvp.exception;
+
+public class NotFoundException extends Exception {
+    public NotFoundException() {
+        super("ERROR: NotFoundException");
+    }
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
