+++ backend/src/main/java/app/linguistai/bmvp/request/QChangePassword.java
@@ -1,0 +1,9 @@
+package app.linguistai.bmvp.request;
+
+import lombok.Data;
+
+@Data
+public class QChangePassword {
+    private String oldPassword;
+    private String newPassword;
+}
