+++ backend/src/main/java/app/linguistai/bmvp/response/RRefreshToken.java
@@ -1,0 +1,10 @@
+package app.linguistai.bmvp.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class RRefreshToken {
+    private String accessToken;
+}
