--- backend/src/main/java/app/linguistai/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package app.linguistai.exception;
-
-public class NotFoundException extends Exception {
-    public NotFoundException() {
-        super("ERROR: NotFoundException");
-    }
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
