--- backend/src/main/java/app/linguistai/request/QChangePassword.java
@@ -1,9 +1,0 @@
-package app.linguistai.request;
-
-import lombok.Data;
-
-@Data
-public class QChangePassword {
-    private String oldPassword;
-    private String newPassword;
-}
