--- backend/src/main/java/app/linguistai/response/RRefreshToken.java
@@ -1,10 +1,0 @@
-package app.linguistai.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class RRefreshToken {
-    private String accessToken;
-}
