--- backend/src/main/java/app/linguistai/security/SpringdocConfig.java
@@ -1,25 +1,0 @@
-package app.linguistai.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-// You can reach Swagger with url: http://localhost:8080/swagger-ui/index.html/
-@OpenAPIDefinition
-@Configuration
-@SecurityScheme(
-        name = "token",
-        type = SecuritySchemeType.HTTP,
-        bearerFormat = "JWT",
-        scheme = "bearer"
-)
-public class SpringdocConfig {
-    @Bean
-    public OpenAPI baseOpenAPI() {
-        return new OpenAPI().info(new Info().title("Spring Doc").version("1.0.0.").description("Spring doc description"));
-    }
-}
