--- backend/src/test/java/app/linguistai/LinguistaiApplicationTests.java
@@ -1,13 +1,0 @@
-package app.linguistai;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LinguistaiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
