+++ backend/src/test/java/app/linguistai/bmvp/LinguistaiApplicationTests.java
@@ -1,0 +1,13 @@
+package app.linguistai.bmvp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LinguistaiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
